Move provenance code into workspace-data
