+++ src/workspace-data/data-table/entity-service/EntitiesContent.js
@@ -9,7 +9,6 @@ import { cohortNotebook, cohortRNotebook, NotebookCreator } from 'src/analysis/u
 import { tools } from 'src/analysis/utils/tool-utils';
 import { requesterPaysWrapper, withRequesterPaysHandler } from 'src/components/bucket-utils';
 import { ButtonSecondary } from 'src/components/common';
-import { DataTableColumnProvenance } from 'src/components/data/data-table-provenance';
 import { icon, spinner } from 'src/components/icons';
 import IGVBrowser from 'src/components/IGVBrowser';
 import IGVFileSelector from 'src/components/IGVFileSelector';
@@ -28,7 +27,6 @@ import wdlLogo from 'src/images/wdl-logo.png';
 import { Ajax } from 'src/libs/ajax';
 import { EntityServiceDataTableProvider } from 'src/libs/ajax/data-table-providers/EntityServiceDataTableProvider';
 import colors from 'src/libs/colors';
-import { useColumnProvenance } from 'src/libs/data-table-provenance';
 import { withErrorReporting } from 'src/libs/error';
 import Events, { extractWorkspaceDetails } from 'src/libs/events';
 import { isFeaturePreviewEnabled } from 'src/libs/feature-previews';
@@ -39,6 +37,8 @@ import * as Style from 'src/libs/style';
 import * as Utils from 'src/libs/utils';
 import * as WorkspaceUtils from 'src/libs/workspace-utils';
 
+import { DataTableColumnProvenance } from '../../provenance/DataTableColumnProvenance';
+import { useColumnProvenance } from '../../provenance/workspace-data-provenance-utils';
 import DataTable from '../shared/DataTable';
 import { AddColumnModal } from './AddColumnModal';
 import { AddEntityModal } from './AddEntityModal';
